Change dtype of im_id to int64 (#148)

+++ ppdet/modeling/architectures/mask_rcnn.py
@@ -274,7 +274,7 @@ class MaskRCNN(object):
         inputs_def = {
             'image':    {'shape': im_shape,  'dtype': 'float32', 'lod_level': 0},
             'im_info':  {'shape': [None, 3], 'dtype': 'float32', 'lod_level': 0},
-            'im_id':    {'shape': [None, 1], 'dtype': 'int32',   'lod_level': 0},
+            'im_id':    {'shape': [None, 1], 'dtype': 'int64',   'lod_level': 0},
             'im_shape': {'shape': [None, 3], 'dtype': 'float32', 'lod_level': 0},
             'gt_bbox':  {'shape': [None, 4], 'dtype': 'float32', 'lod_level': 1},
             'gt_class': {'shape': [None, 1], 'dtype': 'int32',   'lod_level': 1},
